Bug fixes and Associate protection profile skeleton

+++ src/ServiceManagement/RecoveryServices/Commands.RecoveryServices/lib/PSObjects.cs
@@ -347,7 +347,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.SiteRecovery
                             details.ActiveStorageAccount.StorageAccountName;
                         asrProtectionProfile.HyperVReplicaAzureProviderSettingsObject.RecoveryAzureSubscription = 
                             details.ActiveStorageAccount.SubscriptionId;
-                        asrProtectionProfile.HyperVReplicaAzureProviderSettingsObject.ReplicationFrequencyInSeconds = details.ReplicationInterval;
+                        asrProtectionProfile.HyperVReplicaAzureProviderSettingsObject.ReplicationFrequencyInSeconds = (ushort)details.ReplicationInterval;
                         asrProtectionProfile.HyperVReplicaAzureProviderSettingsObject.ReplicationMethod = details.OnlineIrStartTime.HasValue ?
                             Constants.OnlineReplicationMethod : 
                             Constants.OfflineReplicationMethod;
